working variable explorer tests
