+++ src/datascience-ui/react-common/settingsReactSide.ts
@@ -49,6 +49,8 @@ function load() {
             sendSelectionToInteractiveWindow: false,
             markdownRegularExpression: '^(#\\s*%%\\s*\\[markdown\\]|#\\s*\\<markdowncell\\>)',
             codeRegularExpression: '^(#\\s*%%|#\\s*\\<codecell\\>|#\\s*In\\[\\d*?\\]|#\\s*In\\[ \\])',
+            showJupyterVariableExplorer: true,
+            variableExplorerExclude: 'module;builtin_function_or_method',
             extraSettings: {
                 terminalCursor: 'block',
                 theme: 'Default Dark+'
